Arc bible entry schema added
