+++ saltapi/web/schema/bvit.py
@@ -10,9 +10,9 @@ class Bvit(BaseModel):
     id: int = Field(
         ..., title="BVIT id", description="Unique identifier for the BVIT setup"
     )
-    mode: BvitMode = Field(..., title="Instrument mode", description="Instrument mode")
-    filter: BvitFilter = Field(..., title="Filter", description="Filter")
-    neutral_density: BvitNeutralDensity = Field(
+    mode: "BvitMode" = Field(..., title="Instrument mode", description="Instrument mode")
+    filter: "BvitFilter" = Field(..., title="Filter", description="Filter")
+    neutral_density: "BvitNeutralDensity" = Field(
         ..., title="Neutral density", description="Neutral density setting"
     )
     iris_size: float = Field(
