+++ saltapi/web/schema/common.py
@@ -60,6 +60,23 @@ class ExecutedObservation(BaseExecutedObservation):
     )
 
 
+class Lamp(str, Enum):
+    """Calibration lamp(s)."""
+
+    AR = "Ar"
+    AR_AND_THAR = "Ar and ThAr"
+    CUAR = "CuAr"
+    CUAR_AND_XE = "CuAr and Xe"
+    HGAR = "HgAr"
+    HGAR_AND_NE = "HgAr and Ne"
+    NE = "Ne"
+    QTH1 = "QTH1"
+    QTH1_AND_QTH2 = "QTH1 and QTH2"
+    QTH2 = "QTH2"
+    THAR = "ThAr"
+    XE = "Xe"
+
+
 class ObservationProbabilities(BaseModel):
     """Probabilities related to observing."""
